[ENH] budget_activity + mis_builder
