+++ budget_activity/__manifest__.py
@@ -13,6 +13,9 @@
         "security/ir.model.access.csv",
         "report/budget_monitor_report_view.xml",
         "views/res_config_settings_views.xml",
+        "views/mis_report.xml",
+        "views/mis_report_kpi.xml",
+        "views/mis_budget_item.xml",
         "views/account_move_views.xml",
         "views/budget_activity_view.xml",
         "views/budget_activity_group_view.xml",
