+++ budget_activity/models/__init__.py
@@ -5,6 +5,10 @@ from . import account_move
 from . import budget_activity
 from . import budget_activity_group
 from . import account_analytic_line
-from . import base_budget_move
-from . import budget_control
+from . import mis_budget_item
+from . import mis_report
+from . import mis_report_kpi
 from . import mis_report_instance
+from . import base_budget_move
+from . import budget_period
+from . import budget_control
