+++ budget_activity/models/budget_period.py
@@ -10,14 +10,17 @@ class BudgetPeriod(models.Model):
     def _prepare_controls_activity(self, budget_period, doclines):
         controls = set()
         control_analytics = budget_period.control_analytic_account_ids
-        _analytic_field = doclines._budget_analytic_field
-        for i in doclines:
+        budget_moves = doclines.mapped(doclines._budget_field())
+        for i in budget_moves:
             if budget_period.control_all_analytic_accounts:
-                if i[_analytic_field] and i.activity_id:
-                    controls.add((i[_analytic_field].id, i.activity_id.id))
+                if i.analytic_account_id and i.activity_id:
+                    controls.add((i.analytic_account_id.id, i.activity_id.id))
             else:  # Only analtyic in control
-                if i[_analytic_field] in control_analytics and i.activity_id:
-                    controls.add((i.i[_analytic_field].id, i.activity_id.id))
+                if (
+                    i.analytic_account_id in control_analytics
+                    and i.activity_id
+                ):
+                    controls.add((i.analytic_account_id.id, i.activity_id.id))
         # Convert to list of dict, for readibility
         return [{"analytic_id": x[0], "activity_id": x[1]} for x in controls]
 
@@ -49,5 +52,4 @@ class BudgetPeriod(models.Model):
                     )
                     % (instance.report_id.name, activity.display_name)
                 )
-            return kpi
         return super()._get_kpi_by_control_key(instance, kpis, control)
