+++ budget_activity/readme/USAGE.rst
@@ -1,3 +1,16 @@
+To used this module you have to configued Budget Activity and Budget Activity Group first.
+
 #. Go to Budgeting > Configurations > Budget Activity
 #. Create new activity and match with Account > Save
 #. On invoice lines, User can select Activity field and it will auto change account to match with Activity
+
+
+How to configure an Activity Group with a MIS Builder:
+
+#. Go to Invoicing > Configuration > MIS Reporting > MIS Report Templates
+#. Choose you template that you would to see report by activity group
+#. check 'Is Activity'
+#. create new KPI line and select Activity Group field
+#. It will generate expression related with activity group > Save
+
+Then, you can test by create new bills and view your report.
