--- mis_builder_budget_activity_group/README.rst
@@ -1,114 +1,0 @@
-=================================
-MIS Builder Budget Activity Group
-=================================
-
-.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! This file is generated by oca-gen-addon-readme !!
-   !! changes will be overwritten.                   !!
-   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
-.. |badge1| image:: https://img.shields.io/badge/maturity-Alpha-red.png
-    :target: https://odoo-community.org/page/development-status
-    :alt: Alpha
-.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
-    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
-    :alt: License: AGPL-3
-.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fmis--builder-lightgray.png?logo=github
-    :target: https://github.com/OCA/mis-builder/tree/14.0/mis_builder_budget_activity_group
-    :alt: OCA/mis-builder
-.. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/mis-builder-14-0/mis-builder-14-0-mis_builder_budget_activity_group
-    :alt: Translate me on Weblate
-.. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
-    :target: https://runbot.odoo-community.org/runbot/248/14.0
-    :alt: Try me on Runbot
-
-|badge1| |badge2| |badge3| |badge4| |badge5| 
-
-This module help users created budget by Activity Group, and the report will
-show the budgeted values for each Activity Group instead of Account Code.
-
-**Example**
-
-- Relation between Activity Group and Activity
-
-    - Activity Group 1 - Activity1, Activity2
-    - Activity Group 2 - Activity3, Activity4
-
-- Relation between Activity and Account Code
-
-    - Activity1 - 5010000
-    - Activity2 - 5020000
-    - Activity3 - 5020000 (duplicate with Activity2)
-    - Activity4 - 5040000
-
-.. IMPORTANT::
-   This is an alpha version, the data model and design can change at any time without warning.
-   Only for development or testing purpose, do not use in production.
-   `More details on development status <https://odoo-community.org/page/development-status>`_
-
-**Table of contents**
-
-.. contents::
-   :local:
-
-Usage
-=====
-
-To used this module you have to configued Budget Activity and Budget Activity Group first. then you can used this module following:
-
-#. Go to Invoicing > Configuration > MIS Reporting > MIS Report Templates
-#. Choose you template that you would to see report by activity group
-#. check 'Is Activity'
-#. create new KPI line and select Activity Group field
-#. It will generate expression related with activity group > Save
-
-Then, you can test by create new bills and view your report.
-
-Bug Tracker
-===========
-
-Bugs are tracked on `GitHub Issues <https://github.com/OCA/mis-builder/issues>`_.
-In case of trouble, please check there if your issue has already been reported.
-If you spotted it first, help us smashing it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/mis-builder/issues/new?body=module:%20mis_builder_budget_activity_group%0Aversion:%2014.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
-
-Do not contact contributors directly about support or help with technical issues.
-
-Credits
-=======
-
-Authors
-~~~~~~~
-
-* Ecosoft
-
-Contributors
-~~~~~~~~~~~~
-
-* Saran Lim. <saranl@ecosoft.co.th>
-
-Maintainers
-~~~~~~~~~~~
-
-This module is maintained by the OCA.
-
-.. image:: https://odoo-community.org/logo.png
-   :alt: Odoo Community Association
-   :target: https://odoo-community.org
-
-OCA, or the Odoo Community Association, is a nonprofit organization whose
-mission is to support the collaborative development of Odoo features and
-promote its widespread use.
-
-.. |maintainer-Saran440| image:: https://github.com/Saran440.png?size=40px
-    :target: https://github.com/Saran440
-    :alt: Saran440
-
-Current `maintainer <https://odoo-community.org/page/maintainer-role>`__:
-
-|maintainer-Saran440| 
-
-This module is part of the `OCA/mis-builder <https://github.com/OCA/mis-builder/tree/14.0/mis_builder_budget_activity_group>`_ project on GitHub.
-
-You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
