--- mis_builder_budget_activity_group/__init__.py
@@ -1,1 +1,0 @@
-from . import models
