--- mis_builder_budget_activity_group/models/__init__.py
@@ -1,7 +1,0 @@
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
-
-from . import mis_budget_item
-from . import mis_report
-from . import mis_report_kpi
-from . import budget_period
-from . import budget_control
