--- mis_builder_budget_activity_group/readme/CONTRIBUTORS.rst
@@ -1,2 +1,0 @@
-* Saran Lim. <saranl@ecosoft.co.th>
-* Kitti U. <kittiu@ecosoft.co.th>
