--- mis_builder_budget_activity_group/readme/DESCRIPTION.rst
@@ -1,16 +1,0 @@
-This module help users created budget by Activity Group, and the report will
-show the budgeted values for each Activity Group instead of Account Code.
-
-**Example**
-
-- Relation between Activity Group and Activity
-
-    - Activity Group 1 - Activity1, Activity2
-    - Activity Group 2 - Activity3, Activity4
-
-- Relation between Activity and Account Code
-
-    - Activity1 - 5010000
-    - Activity2 - 5020000
-    - Activity3 - 5020000 (duplicate with Activity2)
-    - Activity4 - 5040000
