--- mis_builder_budget_activity_group/readme/USAGE.rst
@@ -1,9 +1,0 @@
-To used this module you have to configued Budget Activity and Budget Activity Group first. then you can used this module following:
-
-#. Go to Invoicing > Configuration > MIS Reporting > MIS Report Templates
-#. Choose you template that you would to see report by activity group
-#. check 'Is Activity'
-#. create new KPI line and select Activity Group field
-#. It will generate expression related with activity group > Save
-
-Then, you can test by create new bills and view your report.
